requirements.txt
Flask==2.3.3
yt-dlp==2025.9.30.72057.dev0
ffmpeg-python==0.2.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7